+++ featuremanagement/_models/_feature_flag.py
@@ -3,8 +3,9 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # -------------------------------------------------------------------------
-from dataclasses import dataclass
-from collections.abc import Mapping
+from ._feature_conditions import FeatureConditions
+from ._allocation import Allocation
+from ._variant_reference import VariantReference
 
 FEATURE_FLAG_ID = "id"
 FEATURE_FLAG_ENABLED = "enabled"
@@ -16,62 +17,6 @@ FEATURE_FILTER_NAME = "name"
 FEATURE_FILTER_REQUIREMENT_TYPE = "requirement_type"
 REQUIREMENT_TYPE_ALL = "All"
 REQUIREMENT_TYPE_ANY = "Any"
-
-
-class FeatureConditions:
-    """
-    Represents the conditions for a feature flag
-    """
-
-    def __init__(self):
-        self._requirement_type = REQUIREMENT_TYPE_ANY
-        self._client_filters = []
-
-    @classmethod
-    def convert_from_json(cls, feature_name, json_value):
-        """
-        Convert a JSON object to FeatureConditions
-
-        :param json: JSON object
-        :type json: dict
-        :return: FeatureConditions
-        :rtype: FeatureConditions
-        """
-        conditions = cls()
-        if json_value is not None and not isinstance(json_value, Mapping):
-            raise AttributeError("Feature flag conditions must be a dictionary")
-        conditions._requirement_type = json_value.get(FEATURE_FILTER_REQUIREMENT_TYPE, REQUIREMENT_TYPE_ANY)
-        conditions._client_filters = json_value.get(FEATURE_FLAG_CLIENT_FILTERS, [])
-        for feature_filter in conditions._client_filters:
-            feature_filter["feature_name"] = feature_name
-        return conditions
-
-    @property
-    def requirement_type(self):
-        """
-        Get the requirement type for the feature flag
-
-        :return: Requirement type
-        :rtype: str
-        """
-        return self._requirement_type
-
-    @property
-    def client_filters(self):
-        """
-        Get the client filters for the feature flag
-
-        :return: Client filters
-        :rtype: list[dict]
-        """
-        return self._client_filters
-
-    def _validate(self, feature_flag_id):
-        if self._requirement_type not in [REQUIREMENT_TYPE_ALL, REQUIREMENT_TYPE_ANY]:
-            raise ValueError(f"Feature flag {feature_flag_id} has invalid requirement type.")
-        for feature_filter in self._client_filters:
-            if feature_filter.get(FEATURE_FILTER_NAME) is None:
-                raise ValueError(f"Feature flag {feature_flag_id} is missing filter name.")
 
 
 class FeatureFlag:
@@ -176,264 +121,6 @@ class FeatureFlag:
         if not isinstance(self._enabled, bool):
             raise ValueError(f"Feature flag {self._id} must be a boolean.")
         self.conditions._validate(self._id)  # pylint: disable=protected-access
-
-
-class Allocation:
-    """
-    Represents an allocation
-    """
-
-    def __init__(self):
-        self._default_when_enabled = None
-        self._default_when_disabled = None
-        self._user = []
-        self._group = []
-        self._percentile = []
-        self._seed = None
-
-    @classmethod
-    def convert_from_json(cls, json):
-        """
-        Convert a JSON object to Allocation
-
-        :param json: JSON object
-        :type json: dict
-        :return: Allocation
-        :rtype: Allocation
-        """
-        if not json:
-            return None
-        allocation = cls()
-        allocation._default_when_enabled = json.get("default_when_enabled")
-        allocation._default_when_disabled = json.get("default_when_disabled")
-        allocation._user = []
-        allocation._group = []
-        allocation._percentile = []
-        if "user" in json:
-            allocations = json.get("user")
-            for user_allocation in allocations:
-                allocation._user.append(UserAllocation(**user_allocation))
-        if "group" in json:
-            allocations = json.get("group")
-            for group_allocation in allocations:
-                allocation._group.append(GroupAllocation(**group_allocation))
-        if "percentile" in json:
-            allocations = json.get("percentile")
-            for percentile_allocation in allocations:
-                allocation._percentile.append(PercentileAllocation.convert_from_json(percentile_allocation))
-        allocation._seed = json.get("seed", "")
-        return allocation
-
-    @property
-    def default_when_enabled(self):
-        """
-        Get the default variant when the feature flag is enabled
-
-        :return: Default variant when the feature flag is enabled
-        :rtype: str
-        """
-        return self._default_when_enabled
-
-    @property
-    def default_when_disabled(self):
-        """
-        Get the default variant when the feature flag is disabled
-
-        :return: Default variant when the feature flag is disabled
-        :rtype: str
-        """
-        return self._default_when_disabled
-
-    @property
-    def user(self):
-        """
-        Get the user allocations
-
-        :return: User allocations
-        :rtype: list[UserAllocation]
-        """
-        return self._user
-
-    @property
-    def group(self):
-        """
-        Get the group allocations
-
-        :return: Group allocations
-        :rtype: list[GroupAllocation]
-        """
-        return self._group
-
-    @property
-    def percentile(self):
-        """
-        Get the percentile allocations
-
-        :return: Percentile allocations
-        :rtype: list[PercentileAllocation]
-        """
-        return self._percentile
-
-    @property
-    def seed(self):
-        """
-        Get the seed for the allocation
-
-        :return: Seed for the allocation
-        :rtype: str
-        """
-        return self._seed
-
-
-@dataclass
-class UserAllocation:
-    """
-    Represents a user allocation
-    """
-
-    variant: str
-    users: list
-
-
-@dataclass
-class GroupAllocation:
-    """
-    Represents a group allocation
-    """
-
-    variant: str
-    groups: list
-
-
-class PercentileAllocation:
-    """
-    Represents a percentile allocation
-    """
-
-    def __init__(self):
-        self._variant = None
-        self._percentile_from = None
-        self._percentile_to = None
-
-    @classmethod
-    def convert_from_json(cls, json):
-        """
-        Convert a JSON object to PercentileAllocation
-
-        :param json: JSON object
-        :type json: dict
-        :return: PercentileAllocation
-        :rtype: PercentileAllocation
-        """
-        if not json:
-            return None
-        user_allocation = cls()
-        user_allocation._variant = json.get("variant")
-        user_allocation._percentile_from = json.get("from")
-        user_allocation._percentile_to = json.get("to")
-        return user_allocation
-
-    @property
-    def variant(self):
-        """
-        Get the variant for the allocation
-
-        :return: Variant for the allocation
-        :rtype: str
-        """
-        return self._variant
-
-    @property
-    def percentile_from(self):
-        """
-        Get the starting percentile for the allocation
-
-        :return: Starting percentile for the allocation
-        :rtype: int
-        """
-        return self._percentile_from
-
-    @property
-    def percentile_to(self):
-        """
-        Get the ending percentile for the allocation
-
-        :return: Ending percentile for the allocation
-        :rtype: int
-        """
-        return self._percentile_to
-
-
-@dataclass
-class VariantReference:
-    """
-    Represents a variant reference
-    """
-
-    def __init__(self):
-        self._name = None
-        self._configuration_value = None
-        self._configuration_reference = None
-        self._status_override = None
-
-    @classmethod
-    def convert_from_json(cls, json):
-        """
-        Convert a JSON object to VariantReference
-
-        :param json: JSON object
-        :type json: dict
-        :return: VariantReference
-        :rtype: VariantReference
-        """
-        if not json:
-            return None
-        variant_reference = cls()
-        variant_reference._name = json.get("name")
-        variant_reference._configuration_value = json.get("configuration_value")
-        variant_reference._configuration_reference = json.get("configuration_reference")
-        variant_reference._status_override = json.get("status_override", None)
-        return variant_reference
-
-    @property
-    def name(self):
-        """
-        Get the name of the variant
-
-        :return: Name of the variant
-        :rtype: str
-        """
-        return self._name
-
-    @property
-    def configuration_value(self):
-        """
-        Get the configuration value for the variant
-
-        :return: Configuration value for the variant
-        :rtype: str
-        """
-        return self._configuration_value
-
-    @property
-    def configuration_reference(self):
-        """
-        Get the configuration reference for the variant
-
-        :return: Configuration reference for the variant
-        :rtype: str
-        """
-        return self._configuration_reference
-
-    @property
-    def status_override(self):
-        """
-        Get the status override for the variant
-
-        :return: Status override for the variant
-        :rtype: str
-        """
-        return self._status_override
 
 
 def _convert_boolean_value(enabled):
